Apply second-audit fixes; sync docs (timing table unchanged)
Code-auditor (gpt-5.5 xhigh) triage; applied the real, low-risk findings and
re-validated (25 tests pass; inference table reproduced with no value change):

- vjepa2_ac_infer_test.py: default --dtype bf16 (bare command matched fp32 before,
  contradicting the documented bf16 requirement).
- franka_droid_env.py: capture_goal_image settles the gripper with a short physics
  rollout (state saved/restored) so open vs closed goal images differ; mj_forward
  alone does not move the driven Robotiq fingers. Added a regression test.
- franka_build.py: anchor DEFAULT_MENAGERIE to the repo root via __file__ (was
  cwd-relative, so FrankaDroidEnv() failed outside the repo root).
- download_checkpoints.py: verify streamed size vs Content-Length; keep .part on
  truncation.
- mujoco_scene.py: default camera wrist_cam -> scene_cam.
- Docs: correct the CEM action bound (per-axis box / L-inf, not an L1 ball) and flag
  the env's L2 clamp; refresh the stale architecture.md V-JEPA section and plan.md
  checklist; wrapper loader points to the local checkpoint, not torch.hub.

Deferred to the experiment stage (real, not defects): unified Franka+vial+holder
scene, pinned vendored/HF revisions, broader test coverage.

+++ scripts/vjepa2_ac_infer_test.py
@@ -122,7 +122,7 @@ def main() -> None:
 
     parser = argparse.ArgumentParser(description=__doc__,
                                      formatter_class=argparse.RawDescriptionHelpFormatter)
-    parser.add_argument("--dtype", choices=["fp32", "bf16"], default="fp32")
+    parser.add_argument("--dtype", choices=["fp32", "bf16"], default="bf16")
     parser.add_argument("--device", default="cuda")
     parser.add_argument("--chunk", type=int, default=200,
                         help="predictor sub-batch size over CEM samples (0 = whole batch); "
